Late night update, quite a few changes
Added level selection, still working on the levels. A bug has appeared
on sliding the levels, only one of them seem to move right now, will
think about it and fix it later on. First sleep. Fixed and tidied up
some of the code on quite a few classes.

Added the guiLevels class for the level selection. Quite the work today